+++ xrpl/models/transactions/__init__.py
@@ -4,7 +4,11 @@ Model objects for specific `types of Transactions
 """
 
 from xrpl.models.transactions.account_delete import AccountDelete
-from xrpl.models.transactions.account_set import AccountSet, AccountSetFlag
+from xrpl.models.transactions.account_set import (
+    AccountSet,
+    AccountSetFlag,
+    AccountSetFlagInterface,
+)
 from xrpl.models.transactions.check_cancel import CheckCancel
 from xrpl.models.transactions.check_cash import CheckCash
 from xrpl.models.transactions.check_create import CheckCreate
@@ -18,14 +22,24 @@ from xrpl.models.transactions.nftoken_cancel_offer import NFTokenCancelOffer
 from xrpl.models.transactions.nftoken_create_offer import (
     NFTokenCreateOffer,
     NFTokenCreateOfferFlag,
+    NFTokenCreateOfferFlagInterface,
 )
-from xrpl.models.transactions.nftoken_mint import NFTokenMint, NFTokenMintFlag
+from xrpl.models.transactions.nftoken_mint import (
+    NFTokenMint,
+    NFTokenMintFlag,
+    NFTokenMintFlagInterface,
+)
 from xrpl.models.transactions.offer_cancel import OfferCancel
-from xrpl.models.transactions.offer_create import OfferCreate, OfferCreateFlag
-from xrpl.models.transactions.payment import Payment, PaymentFlag
+from xrpl.models.transactions.offer_create import (
+    OfferCreate,
+    OfferCreateFlag,
+    OfferCreateFlagInterface,
+)
+from xrpl.models.transactions.payment import Payment, PaymentFlag, PaymentFlagInterface
 from xrpl.models.transactions.payment_channel_claim import (
     PaymentChannelClaim,
     PaymentChannelClaimFlag,
+    PaymentChannelClaimFlagInterface,
 )
 from xrpl.models.transactions.payment_channel_create import PaymentChannelCreate
 from xrpl.models.transactions.payment_channel_fund import PaymentChannelFund
@@ -33,12 +47,17 @@ from xrpl.models.transactions.set_regular_key import SetRegularKey
 from xrpl.models.transactions.signer_list_set import SignerEntry, SignerListSet
 from xrpl.models.transactions.ticket_create import TicketCreate
 from xrpl.models.transactions.transaction import Memo, Signer, Transaction
-from xrpl.models.transactions.trust_set import TrustSet, TrustSetFlag
+from xrpl.models.transactions.trust_set import (
+    TrustSet,
+    TrustSetFlag,
+    TrustSetFlagInterface,
+)
 
 __all__ = [
     "AccountDelete",
     "AccountSet",
     "AccountSetFlag",
+    "AccountSetFlagInterface",
     "CheckCancel",
     "CheckCash",
     "CheckCreate",
@@ -52,15 +71,20 @@ __all__ = [
     "NFTokenCancelOffer",
     "NFTokenCreateOffer",
     "NFTokenCreateOfferFlag",
+    "NFTokenCreateOfferFlagInterface",
     "NFTokenMint",
     "NFTokenMintFlag",
+    "NFTokenMintFlagInterface",
     "OfferCancel",
     "OfferCreate",
     "OfferCreateFlag",
+    "OfferCreateFlagInterface",
     "Payment",
     "PaymentFlag",
+    "PaymentFlagInterface",
     "PaymentChannelClaim",
     "PaymentChannelClaimFlag",
+    "PaymentChannelClaimFlagInterface",
     "PaymentChannelCreate",
     "PaymentChannelFund",
     "SetRegularKey",
@@ -71,4 +95,5 @@ __all__ = [
     "Transaction",
     "TrustSet",
     "TrustSetFlag",
+    "TrustSetFlagInterface",
 ]