+++ xrpl/models/transactions/pseudo_transactions/__init__.py
@@ -5,8 +5,15 @@ Model objects for specific `types of Pseudo-Transactions
 from xrpl.models.transactions.pseudo_transactions.enable_amendment import (
     EnableAmendment,
     EnableAmendmentFlag,
+    EnableAmendmentFlagInterface,
 )
 from xrpl.models.transactions.pseudo_transactions.set_fee import SetFee
 from xrpl.models.transactions.pseudo_transactions.unl_modify import UNLModify
 
-__all__ = ["EnableAmendment", "EnableAmendmentFlag", "SetFee", "UNLModify"]
+__all__ = [
+    "EnableAmendment",
+    "EnableAmendmentFlag",
+    "SetFee",
+    "UNLModify",
+    "EnableAmendmentFlagInterface",
+]